refactor(cli): add commands module scaffold + shared helpers (common)
Introduce crate::commands as the new home for the CLI handlers:

- prelude.rs: the shared import surface every handler pulls in via
  `use crate::commands::prelude::*` — clap defs (crate::cli), the common
  helpers below, the few main.rs-resident items handlers call, and the
  std/external symbols used by short name. Command groups are re-exported
  here as they are split out so cross-group calls resolve without
  per-call-site imports.
- common.rs: 42 cross-cutting helpers moved out of main.rs — daemon
  discovery + HTTP client, JSON-over-HTTP, home/config resolution,
  formatting, and small filesystem/clipboard utils — plus the
  DaemonConfigContext struct.

main.rs now imports the prelude; install_ctrlc_handler is pub(crate) for
handler reuse. Pure code move, no behavior change.
